+++ cpp/demo/include/dog.h
@@ -1,0 +1,10 @@
+#ifndef DOG_H_
+#define DOG_H_
+#include <iostream>
+
+class Dog
+{
+public:
+    u_int8_t age(uint8_t param);
+};
+#endif // !DOG_H_