+++ cpp/demo/src/cat.c
@@ -1,0 +1,10 @@
+#include <stdio.h>
+
+int main()
+{
+    unsigned short age = 12;
+
+    printf("age is %d\n", age);
+
+    return 0;
+}